added supabase.ts and moved colors.ts
